Move remote user helpers to separate file
