+++ python/nav/web/auth/__init__.py
@@ -21,8 +21,6 @@ The "*Middleware" is Django-specific.
 
 from datetime import datetime
 import logging
-from os.path import join
-import os
 
 from urllib import parse
 
@@ -36,27 +34,17 @@ try:
 except ImportError:  # Django <= 1.9
     MiddlewareMixin = object
 
-try:
-    # Python 3.6+
-    import secrets
-
-    def fake_password(length):
-        return secrets.token_urlsafe(length)
-
-except ImportError:
-    from random import choice
-    import string
-
-    def fake_password(length):
-        symbols = string.ascii_letters + string.punctuation + string.digits
-        return u"".join(choice(symbols) for i in range(length))
-
 
 from nav.auditlog.models import LogEntry
-from nav.config import NAVConfigParser
 from nav.django.utils import is_admin, get_account
 from nav.models.profiles import Account, AccountGroup
 from nav.web import ldapauth
+from nav.web.auth.remote_user import (
+    authenticate_remote_user,
+    get_remote_loginurl,
+    get_remote_logouturl,
+    get_remote_username,
+)
 
 
 _logger = logging.getLogger(__name__)
@@ -73,21 +61,6 @@ LOGIN_URL = '/index/login/'
 # If the entire site is protected via remote_user, this link must be outside
 # that protection!
 LOGOUT_URL = '/index/logout/'
-
-
-class RemoteUserConfigParser(NAVConfigParser):
-    DEFAULT_CONFIG_FILES = [join('webfront', 'webfront.conf')]
-    DEFAULT_CONFIG = u"""
-[remote-user]
-enabled=no
-login-url=
-logout-url=
-varname=REMOTE_USER
-workaround=none
-"""
-
-
-_config = RemoteUserConfigParser()
 
 
 def _set_account(request, account):
@@ -166,48 +139,6 @@ def _handle_ldap_admin_status(ldap_user, nav_account):
             nav_account.groups.remove(admin_group)
 
 
-def authenticate_remote_user(request):
-    """Authenticate username from http header REMOTE_USER
-
-    Returns:
-
-    :return: If the user was authenticated, an account.
-             If the user was blocked from logging in, False.
-             Otherwise, None.
-    :rtype: Account, False, None
-    """
-    username = get_remote_username(request)
-    if not username:
-        return None
-
-    # We now have a username-ish
-
-    try:
-        account = Account.objects.get(login=username)
-    except Account.DoesNotExist:
-        # Store the remote user in the database and return the new account
-        account = Account(login=username, name=username, ext_sync='REMOTE_USER')
-        account.set_password(fake_password(32))
-        account.save()
-        _logger.info("Created user %s from header REMOTE_USER", account.login)
-        template = 'Account "{actor}" created due to REMOTE_USER HTTP header'
-        LogEntry.add_log_entry(
-            account, 'create-account', template=template, subsystem='auth'
-        )
-        return account
-
-    # Bail out! Potentially evil user
-    if account.locked:
-        _logger.info("Locked user %s tried to log in", account.login)
-        template = 'Account "{actor}" was prevented from logging in: blocked'
-        LogEntry.add_log_entry(
-            account, 'login-prevent', template=template, subsystem='auth'
-        )
-        return False
-
-    return account
-
-
 def get_login_url(request):
     """Calculate which login_url to use"""
     path = parse.quote(request.get_full_path())
@@ -225,104 +156,6 @@ def get_logout_url(request):
     if remote_logouturl and remote_logouturl.endswith('='):
         remote_logouturl += request.build_absolute_uri(LOGOUT_URL)
     return remote_logouturl if remote_logouturl else LOGOUT_URL
-
-
-def get_remote_loginurl(request):
-    """Return a url (if set) to log in to/via a remote service
-
-    :return: Either a string with an url, or None.
-    :rtype: str, None
-    """
-    return get_remote_url(request, 'login-url')
-
-
-def get_remote_logouturl(request):
-    """Return a url (if set) to log out to/via a remote service
-
-    :return: Either a string with an url, or None.
-    :rtype: str, None
-    """
-    return get_remote_url(request, 'logout-url')
-
-
-def get_remote_url(request, urltype):
-    """Return a url (if set) to a remote service for REMOTE_USER purposes
-
-    :return: Either a string with an url, or None.
-    :rtype: str, None
-    """
-    remote_url = None
-    try:
-        if not _config.getboolean('remote-user', 'enabled'):
-            return None
-        remote_url = _config.get('remote-user', urltype)
-    except ValueError:
-        return None
-    if remote_url:
-        nexthop = request.build_absolute_uri(request.get_full_path())
-        remote_url = remote_url.format(nexthop)
-    return remote_url
-
-
-def get_remote_username(request):
-    """Return the username in REMOTE_USER if set and enabled
-
-    :return: The username in REMOTE_USER if any, or None.
-    :rtype: str, None
-    """
-    try:
-        if not _config.getboolean('remote-user', 'enabled'):
-            return None
-    except ValueError:
-        return None
-
-    if not request:
-        return None
-
-    workaround = 'none'
-    try:
-        workaround_config = _config.get('remote-user', 'workaround')
-    except ValueError:
-        pass
-    else:
-        if workaround_config in REMOTE_USER_WORKAROUNDS:
-            workaround = workaround_config
-
-    username = REMOTE_USER_WORKAROUNDS[workaround](request)
-
-    if not username:
-        return None
-
-    return username
-
-
-def _get_remote_user_varname():
-    varname = 'REMOTE_USER'
-    try:
-        varname = _config.get('remote-user', 'varname')
-    except ValueError:
-        pass
-    return varname
-
-
-def _workaround_default(request):
-    varname = _get_remote_user_varname()
-    username = request.META.get(varname, '').strip()
-    return username
-
-
-def _workaround_feide_oidc(request):
-    varname = _get_remote_user_varname()
-    username = request.META.get(varname, '').strip()
-    if ':' in username:
-        username = username.split(':', 1)[1]
-    return username
-
-
-REMOTE_USER_WORKAROUNDS = {
-    'none': _workaround_default,
-    'feide-oidc': _workaround_feide_oidc,
-}
 
 
 # Middleware
